requirements-gpu.txt
jupyter==1.0.0
numpy==1.14.1
pandas==0.22.0
requests==2.18.4
scikit_learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
tensorflow-gpu==1.6.0